WASM64 WIP

+++ onnxruntime/core/providers/js/js_kernel.h
@@ -90,20 +90,20 @@ class JsKernel : public OpKernel {
     for (int i = 0; i < context->InputCount(); i++) {
       temp_data_size += sizeof(size_t) * (3 + context->Input<Tensor>(i)->Shape().NumDimensions());
     }
-    uint32_t* p_serialized_kernel_context = reinterpret_cast<uint32_t*>(alloc->Alloc(temp_data_size));
+    size_t *p_serialized_kernel_context = reinterpret_cast<size_t*>(alloc->Alloc(temp_data_size));
     if (p_serialized_kernel_context == nullptr) {
       return nullptr;
     }
 
-    p_serialized_kernel_context[0] = reinterpret_cast<uint32_t>(context);
-    p_serialized_kernel_context[1] = static_cast<uint32_t>(context->InputCount());
+    p_serialized_kernel_context[0] = reinterpret_cast<size_t>(context);
+    p_serialized_kernel_context[1] = static_cast<size_t>(context->InputCount());
     size_t index = 2;
     for (int i = 0; i < context->InputCount(); i++) {
-      p_serialized_kernel_context[index++] = static_cast<uint32_t>(context->Input<Tensor>(i)->GetElementType());
-      p_serialized_kernel_context[index++] = reinterpret_cast<uint32_t>(context->Input<Tensor>(i)->DataRaw());
-      p_serialized_kernel_context[index++] = static_cast<uint32_t>(context->Input<Tensor>(i)->Shape().NumDimensions());
+      p_serialized_kernel_context[index++] = static_cast<size_t>(context->Input<Tensor>(i)->GetElementType());
+      p_serialized_kernel_context[index++] = reinterpret_cast<size_t>(context->Input<Tensor>(i)->DataRaw());
+      p_serialized_kernel_context[index++] = static_cast<size_t>(context->Input<Tensor>(i)->Shape().NumDimensions());
       for (size_t d = 0; d < context->Input<Tensor>(i)->Shape().NumDimensions(); d++) {
-        p_serialized_kernel_context[index++] = static_cast<uint32_t>(context->Input<Tensor>(i)->Shape()[d]);
+        p_serialized_kernel_context[index++] = static_cast<size_t>(context->Input<Tensor>(i)->Shape()[d]);
       }
     }
