+++ onnxruntime/core/providers/js/operators/transpose.h
@@ -24,10 +24,10 @@ class Transpose final : public JsKernel, public TransposeBase {
                                  "perm" : $1 ? Array.from(HEAP32.subarray($2, $2 + $1)) : []
                                }),
                                // $1: length of attribute "perm" (int32[])
-                               gsl::narrow_cast<int32_t>(perm_specified_ ? perm_.size() : 0),
+                               gsl::narrow_cast<size_t>(perm_specified_ ? perm_.size() : 0),
                                // $2: index to HEAP32 of the first int32 element. calculated from right shift memory
                                //     address by 2
-                               reinterpret_cast<int32_t>(perm_specified_ && !perm.empty() ? perm.data() : nullptr) >> 2);
+                               reinterpret_cast<size_t>(perm_specified_ && !perm.empty() ? perm.data() : nullptr) >> 2);
   }
 };
